Enable modules before starting Apache
If Puppet decides to order resources in such a way that Apache is
started before the modules the vhost is configured to utilize are
enabled, the service will fail complaining of a syntax error/missing
module. Set Apache's modules to instead be enabled before Apache is
started to avoid this failure scenario.

Change-Id: Ida3e963d9044bbb5e0ec4c9e465a104e7d97fa2b